[!!!][TASK] Remove per-user IP locking for backend users
The IP-locking functionality is disabled in TYPO3 v10 by default due to
IPv4/IPv6 hopping ("Happy eyeballs"), however the per-Backend-User
option is now removed, as it effectively needs custom enablement.

IF IP locking is needed for the backend, this is now configurable
only on a per-installation basis. If needed, it is recommended to
implement this feature, which was disabled by default
(which depends on a feature, which is disabled by default) as
a custom extension.

Resolves: #92559
Releases: master
Change-Id: I9b9b70dab4d3240bd72d8b3979140fc1a5a46811
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66093
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Benni Mack <benni@typo3.org>