[TASK] Remove unused log argument in DataHandler
Since TYPO3 v10 the history is connected via
the correlationId in DataHandler and the setHistory()
method does not need the third argument anymore.

This change removes the third argument, and removes
the only two usages in Workspaces.

The method is marked as internal, and the argument
has no value. For this reason, we can remove the
method argument.

Resolves: #92569
Releases: master
Change-Id: I298e24b81336dbddfd477f0881f752f99300a894
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66144
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>