[BUGFIX] Make recordlist collapse button accessible
A <button> tag is now being used instead of an <a> tag
as collapsing / expanding the tables is an interaction on the
same page rather than opening a new one (<a> tag had empty
`href` attribute).

The button now also provides a proper aria-label with a
descriptive text and gets the aria-expandend state
initially set.

Furthermore the button now uses common styling to be
directly recognized as such.

Resolves: #92657
Releases: master
Change-Id: I7d98720f49b549bd74af1632c1fdd9fb9accbdba
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66244
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Tested-by: Daniel Haupt <mail@danielhaupt.de>
Tested-by: Jens Ulrich <jens.ulrich@snk.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Reviewed-by: Daniel Haupt <mail@danielhaupt.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>