[TASK] Build extbase requests early in Bootstrap
This patch introduces a couple of changes to the extbase request
creation and handling process.

The request builder is used during bootstrap to create an extbase
request object. This enables building the extbase request from the
incoming server request in the future without passing the server
request through a layer of services.

Further, it enables a more sensible API for the request handlers
whose methods canHandleRequest() and handleRequest() did not receive
the extbase request as arguments for evaluation in the past.

The idea is to have the extbase request and an attached context object
as single source of truth regarding request data and avoid storing
that data elsewhere or accessing it through services during runtime.

Releases: master
Resolves: #92703
Change-Id: I6c9c3fd32e1af32ef3f73aa12073d34c8a63a43d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66281
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>