[TASK] Use Symfony Routing for Backend Routing
This is a pre-cursor for enhancing Backend Routing in general.

This change implements Symfony Routing for Backend
Routes (currently based on the "route" GET/POST argument)
internally to make use of their compiled routes later-on.

This was previously done in a simplified version where only
the path was checked against, and most Symfony features
were not in use, which we still don't use yet, but more can now
be added.

When Backend Routing was added in TYPO3 v7, the API
was heavily influenced by Symfony but we did not have
a lot of experience yet, which we benefitted in TYPO3 v9+
for Frontend Routing.

With this experience (and with PSR-15) we can gradually
move to a more streamlined and faster API.

The proposed solution still works as expected with &route
GET parameters, but opens up the door for flexible Route arguments.

Resolves: #92723
Releases: master
Change-Id: I48475fcb3cc15f99cc102dac36ae15bca9f3032e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66316
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>