[TASK] Always handle linktypes in linkvalidator as list
The TSconfig of linkvalidator contains 'linktypes' which is a
list of link types which should be checked
(e.g. 'external','db' ...).

This list of links can have 2 different formats inside internal code:

['external', 'db']
or

[
   'external' => 1,
   'db' => 1
]

Converting this in various places made the code hard to read and error
prone.

Only the first format is now used, which also makes it possible to
simplify the tests.

Resolves: #92715
Releases: master
Change-Id: I7451f707da15a20b21b5fefe5df50ed455a5b733
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66319
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>