[BUGFIX] Use correct format character in CleanerTaskTest
The CleanerTaskTest creates a DateInterval and formats it
to return the day representation of the difference. Therefore,
the `%d` format character was used. Because `%d` does only
output the number of days not already covered by the month,
the test now failed.

The `%a` format character is now used as it will always output
the total number of days.

Resolves: #92727
Releases: master, 10.4
Change-Id: I2e7dd9af41b54f5e5dbaea39eaaac1724fd29982
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66328
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>