[FEATURE] Add button to create sys_note records directly in modules
System notes are being displayed in various modules.
Previously one however had to always switch to the list
module to add such note.

Therefore, a new button is added to the button bar of the
page, list and info module which allows to directly create
a new sys_note record for the current page. The new button
can be disabled via page TSconfig.

Resolves: #83814
Releases: master
Change-Id: Id7ab3b7a25d14c67a2a1b41753e1eae34eecd69e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66373
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>