[BUGFIX] Fix type error in TCA table wizard for null values
When using renderType textTable for text fields in TCA, the controller
will pass the currently stored value to underlying API. This API always
expects a string. Most of the time those fields might have null as
default in persistence leading to a PHP TypeError.

Therefore a check is added which will fall back to empty string as
proper default for internal API.

Resolves: #92761
Releases: master, 10.4, 9.5
Change-Id: I29ef4076e21a1f93f9b6834afcc163c4041a0909
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66381
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>