[TASK] Remove BootstrapInterface from Extbase
The BootstrapInterface was used in the original
design to allow interchangeability, but was never
flexible enough and much bound to cObj to have
any effect. The interface was marked as internal
in previous TYPO3 versions and is removed.

Analysis throughout Extbase extensions in TER
have shown that e.g. TypoScript references
use the Bootstrap class directly.

Resolves: #92767
Releases: master
Change-Id: I2514af70d760e9affda8f3150560440c3d842337
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66416
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>