[BUGFIX] Show flag for language "All" in the Page module
When switching to the "Language" module in the Backend, the module
will show language flags and titles for content elements. However
if the content element has language "All", neither flag, nor title
will be displayed. This change adds the flag and the title.

The fix is contributed by the University of Basel.

Resolves: #92785
Releases: 9.5
Change-Id: Id26aaf39f1145b3075a4ff32bd263e1c672be4cc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66575
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Dmitry Dulepov <9csxfqr4jy@liamekaens.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Dmitry Dulepov <9csxfqr4jy@liamekaens.com>
Reviewed-by: Benni Mack <benni@typo3.org>