[!!!][TASK] Remove feature security.frontend.keepSessionDataOnLogout
The feature flag "security.frontend.keepSessionDataOnLogout"
was introduced as part of a security bugfix to still enable frontend
users to keep their session data even if they have logged out,
where the session data was transferred and migrated to an
anonymous session.

Since this feature in general is insecure, as people who log off
from a public computer would keep session data on that machine,
the functionality is fully removed.

Resolves: #92807
Releases: master
Change-Id: Ieaebcc33e85e1df6e359a7eae318712896800bca
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66596
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>