[TASK] Public method to retrieve class loader
To bootstrap TYPO3 sub requests, the testing framework needs an
instance of the class loader. There is no good way to retrieve
this object that is immutable for the entire process.

During casual bootstrap, an instance of the class loader is parked
in object ClassLoadingInformation. The patch makes getClassLoader()
public to allow retrieval.

This solution is kinda ugly and should probably change later. The
public method is for now marked @internal and should not be used
by third party code to allow structural changes later.

Resolves: #92883
Releases: master
Change-Id: Ic819e1a5989a74ebf634e1b1090058b7fe93af9e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66708
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>