[FEATURE] AbstractApplication implements PSR-15 RequestHandlerInterface
Frontend, backend and install tool application classes are the
main entry points to create a response from a request object.

From within a HTTP index.php, run() is called, which creates the
request object from globals and feeds that to handle().

But, to retrieve a TYPO3 response from within a different PHP
application, or from within TYPO3 itself (eg. a backend calls a
frontend), the leading application would want to hand over a
given or specially crafted PSR-7 ServerRequest directly.

This is exactly what PSR-15 RequestHandlerInterface is for.

By implementing this inteface in the application classes, the
core increases interoperability significantly and allows to
be easily called by a third party PHP application.

Resolves: #92884
Releases: master
Change-Id: I3047c92de06668db4dd5ef224bafde23f4b8ebd5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66710
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>