[BUGFIX] Extend condition for isOnSymmetricSide
Instead of holding only the uid as a plain value the given databaseRow
array of the symmetric_field can hold an array containing a lot of
information about the related foreign record.
This e.g. applies if type group is used instead of type select.
Now isOnSymmetricSide is set to true in both cases.

Resolves: #92963
Releases: master
Change-Id: Ibc761831a757915f454f08d64e4c13908e1cbb7d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66955
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>