[!!!][TASK] Separate login failure in authentication process
The flag AbstractUserAuthentication->loginFailure
is removed. The current plans are to clean up the
AbstractUserAuthentication class, and build separate
concerns as much as possible.

The main "monster" method "checkAuthentication" is now
cleaned up, and the public property "loginFailure" is removed,
as this is handled by a local variable now.

Resolves: #92989
Releases: master
Change-Id: I377baa94d6681fab7de04a74991254156235af52
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66999
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>