[BUGFIX] Respect none accessible workspace pages in tree list
If a versioned page is inaccessible due to versionOL, the row will be
converted to a value of "false". This can be the case for deleted or hidden pages.
In those cases, the tree should not include this page.
Keeping the page would furthermore result in a type error, as later
calls expect $row to be an array.

Resolves: #93009
Releases: master, 10.4
Change-Id: I43033e6fe3b403ffc2eebb2cc7080b988425917e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67016
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>