[BUGFIX] Fix negated isLoggedInBackendUserRequired() method
The method returned the opposite of what the function name
defines. It checks whether the current backend route
is a public route, and in that case it returned true. But if
a public route is requested, then a backend user is actually
*not* required.

The method was then used inverted, which contervailed this mistake.
Therefore there was no functional bug here, just a naming bug.

Therefore we now invert the result of this method.

Releases: master
Resolves: #93021
Change-Id: I4ae585eb9259360cb3975df6654640d18ec45932
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67048
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>