[FEATURE] Store Flash Messages as JsonSerializable
Flash Messages which should be stored in sessions,
are now stored as stringified JsonSerializable instead of
their objects themselves.

The huge advantages over the previous functionality
is that the FlashMessage objects are not instantiated during the process
when a session is loaded, but actually only when the FlashMessages
are needed, bringing a bit of performance for Backend functionality.

Resolves: #93063
Releases: master
Change-Id: Ie2da3b3d0a3fbb14416fd5743b3109221d30d28d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67098
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>