[!!!][TASK] Protect RelationHandler internals
Class RelationHandler is a confusing mess with tons
of state spread over 28 class properties and 35
consuming methods. Nearly everything is public and
of course far from bug free.

The patch protects a series of properties and methods
that handle internal state and do not need to be
accessed from outside. For instance, the same state
created by the 'read' methods is done by start().

This is a starter patch to allow refactorings
towards a better manageable codebase.

Change-Id: I8aa45fcd0306209fcc298972ad850449d2c4f125
Resolves: #93080
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67138
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>