[!!!][TASK] Simplify host information for language packs
The host which contains the language zip files is now hardcoded into the
`LanguagePackService` instead of fetching an outdated information from
typo3.org and overriding it on the fly again.

Persisting and checking the registry is not needed anymore.
If the host needs to be overridden, the provided event
`ModifyLanguagePackRemoteBaseUrlEvent` can be used.

Resolves: #93081
Releases: master
Change-Id: I7acced228745f8ae428ff6474800d4e15c2e4ca7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67141
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>