[BUGFIX] Fix requirejs-loader config and error handling
Adapt the fetch-configuration URL query string, to account for the
backend url change in #93048, in order to use a questionmark
to concatenate the query string instead of an ampersand.
But allow for both, base URLs with and without question mark,
as frontend context via eID would still require an ampersand
for concatenation.

Also fix a error callback name clash and properly propagate
fetchConfiguration errors through the requirejs error handling,
for requirejs errbacks to be invoked.

Resolves: #93227
Related: #93048
Releases: master, 10.4
Change-Id: If843db529e19e6af2681871a2d6307d67acbea56
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67307
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>