[BUGFIX] Remove link from login logo
The login logo is linked only for technical reasons of a security check.
It is not meant to be used by humans but is placed prominently enough
to be an accessibility issue.

Remove the link around the logo and create a new one hidden from the
user interface.

Resolves: #93172
Releases: master, 10.4
Change-Id: Id1edc5dd0e059a8700e4bb6c0e09e3f443b6f9fb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67252
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>