[BUGFIX] FormEngine: Fix several validation issues in complex field types
The FormEngine field type `selectMutlipleSideBySide` is very special, as
it's a combination of four input and select fields. With the refactoring
done in #87324 an issue in the validation has been introduced which
caused the validation of the wrong field which contains no validation
rules. This patch uses the correct field to obtain the validation rules
from, now.

Another issue is that values of `selectCheckBox` never have been
validated as the necessary validation rules were missing. The rules are
now appended to the outermost div container. In the same run, minor
styling and markup changes were necessary.

Resolves: #93191
Related: #87324
Releases: master, 10.4
Change-Id: I7195213a768fa4814eb3ecd81953a500c77ede5d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67356
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>