[BUGFIX] Reduce ses_id field length
The patch for #93131 increased the db size of field ses_id
in tables fe_sessions and be_sessions from 32 to 255 chars.
The field is the tables primary key.
With mysql/mariadb, if the table is created on an older db
engine version that creates innodb tables with COMPACT
row format using utf8_mb4 charset, the key length now
exceeds the COMPACT maximum index length restrition of 767
bytes: 4*255+1 = 1021 bytes
Since we don't strictly need a varchar(255), the patch
reduces the field to 190, which does not collide with
maximum key length restrictions: 4*190+1 = 761 bytes

Change-Id: Ifc344a68a49ce0b863109cbfe8a7d27c0179c0cf
Resolves: #93150
Related: #93131
Releases: master, 10.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67336
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>