[BUGFIX] Add default workspace for dummy user
The backend usergroup compare view requires to fetch the
group data of the groups to compare. Since this operation
depends on a user array, UserInformationService creates
such dummy user.

The dummy user array now also contains the default workspace
id `-99` to prevent triggering a database update which would
result in creating a sys log entry. This would fail since the user,
the database update should be performed on, is the dummy
user which is not fully set up.

Resolves: #93233
Releases: master, 10.4
Change-Id: I61870acbab133d9f3f1b80899b50662dca2bc9bf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67352
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>