[BUGFIX] Fix RenameFile and FileDelete in EXT:filelist
The RenameFile JavaScript handles renaming of files
and folders. If renaming a folder, some input fields are
not available in the form and must therefore be checked
for existence, before accessing their value.

Furthermore does the file delete modal now display the
modal content again. The corresponding data attribute
was renamed in the bootstrap v5 migration.

Resolves: #93277
Releases: master
Change-Id: I4ba6bc625a152d7e97ba6c44612a17b782799d31
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67438
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>