[FEATURE] Synchronize folder relations after rename
To prevent broken records which reference a renamed
folder, the AfterFolderRenamedEvent is now used to
synchronize the corresponding relations as soon as
such action took place.

Furthermore does the AfterFolderRenamedEvent now
contain the source folder, which is necessary to
find those relations.

A flash message informs the user about the
synchronization when the action was performed
in the backend, e.g. in the file module.

Resolves: #78036
Resolves: #83054
Releases: master
Change-Id: I9cc8bfbb5c8506c22dc4377fd3041b1113864182
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67441
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>