[BUGFIX] Allow backend login with invalid mailer notification
When a TYPO3 installation has `warning_email_addr` set,
but the mailer configuration is wrong (e.g. cannot connect
to mail server), then no login to backend or install tool
is possible anymore.

This change gracefully catches any exceptions,
and logs the information that an email could not
be sent anymore.

Resolves: #93297
Releases: master, 10.4
Change-Id: I1b658740a64cf1d29b7842f0e2b49d69ce4d5e25
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67597
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>