[FEATURE] Use dropdown for secondary action items
This change replaces the "collapse-horizontal" functionality
of the extended view icons into a dropdown from Bootstrap,
allowing users to understand the icons better because the
text is now also written next to the icon (similar to
the Context Menu).

The horizontal collapsing is removed as a UX improvement
to
a) give consistency around other areas in TYPO3 Backend
b) was intended as a workaround to solve the Extended View rendering and to avoid button-hell
c) avoid confusion around editors when buttons are visually moving horizontally
d) the Extended View feature was unclear for editors when to enable it

In addition the "View" link is now moved into the secondary button group.

The Extended View checkbox is therefore removed as all secondary
options are always available.

Same functionality is also replaced in the File List module to use
the "..." icon to show that there are more actions available.

Side note: The dropdown needs a position-static due to
overflow-x: auto - see https://github.com/twbs/bootstrap/issues/29313#issuecomment-758412385

Resolves: #94390
Releases: master
Change-Id: I925aa7690068fbde3f3aa5217f3f5629a2a890d1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67458
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>