[BUGFIX] Never fetch content with delete placeholder
The content fetcher did not consider delete placeholders.
This task was done using a method in GridColumnItem which
was used in the templates directly. Because defLangBinding
is more complex, three additional calls to this method would
be needed.

Since such records shouldn't even be passed to the view at all,
this check is now directly added into the content fetcher. This
means that no elements with a delete placeholder are fetched
anymore, making the getter and the calls in the templates
superfluous. Therefore, the getter along with its single usage
is removed.

Resolves: #93332
Releases: master, 10.4
Change-Id: I5f214bc8f94ffe571289f7c62849590ca9931ea4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67492
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>