[BUGFIX] Properly position ContextMenu again
With #89496 the ContextMenu got keyboard support
added. However, the underlying code does still
use only a click event to trigger the ContextMenu,
which is now also manually triggered for keypress
events. It was therefore not possible to differ
between the three possible ways, the ContextMenu
can be used:

1. Mouse click only
2. Mixed
2.1 Click on ContextMenu and navigate via keyboard
2.2 Access ContextMenu via keyboard and navigate via click
3. Keyboard only

Therefore, we now check the mouse position. If it's NULL,
the ContextMenu was openend via keyboard. Furthermore, we
use a boolean flag, which helps the getPosition method to
determine the correct position in mixed mode. Using mouse
click will now behave like before. Using the keyboard will
now always use the boundingRect for positioning.

Resolves: #93362
Releases: master
Change-Id: I1e2be756b85d0a2286524f0bdc9cca76208935d1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67510
Tested-by: Richard Haeser <richard@richardhaeser.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Benni Mack <benni@typo3.org>