[FEATURE] Enable resizing & toggling of navigation area
The button in the topbar for switching of the navigation area
(page tree for example) is removed in favor of:

1) a button to switch the navigation frame on / off layered above the nav container
2) a resizable handler on the right to finally resize the navigation tree again

The feature now keeps the previously selected
size in the users' "uc" configuration. When a user resizes
the pagetree, it is kept during reload and next login.

It is possible to resize (drag&drop) via the mouse or with touch events.

Kudos to Benjamin Kott for his first draft in 2019 for
this feature, where some parts of this implementation are taken from.

Resolves: #78760
Resolves: #89355
Releases: master
Change-Id: I4ead734fe2a4e7ce93191eeb03c3312ce7d94643
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67511
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>