[TASK] runTests.sh: Add phpstan and docBlockCheck
Allows:
* Build/Scripts/runTests.sh -s docBlockCheck
* Build/Scripts/runTests.sh -s phpstan

These two have been added to bamboo with #89023, but not to
runTests.sh.

phpstan is tuned a bit to create it's cache in .cache rather
than ephermal /tmp to survive between host reboots.
Additionally the number of max CPU's is restricted in phpstan
to be a better neighbor on local dev machines and on CI.

Resolves: #93367
Related: #89023
Releases: master, 10.4
Change-Id: I5672370b51205266ee46888837f151de8e809edf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67534
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>