[TASK] runTests.sh: Add install on mysql and grunt clean check
Allow:
* Build/Scripts/runTests.sh -s install -d mysql
* Build/Scripts/runTests.sh -s checkGruntClean

The acceptance tests of the install procedure were
missing an option for mysql as opposed to mariadb.

The grunt clean operation compiles everything (js, css, ..)
and verifies no files are changed afterwards. This is mostly
used in CI to ensure patches are pushed using 'current'
yarn/npm dependencies. Note this calls a 'git add', so it
might lead to unexpected git status results if executed
locally.

Change-Id: I21b235b2d264465166b5497e4b2bec25b45e7996
Resolves: #93369
Releases: master, 10.4, 9.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67538
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>