[DOCS] Added condition to snippet for social media fallback images
It is possible to have multiple og:image meta tags by the specs. By
adding the condition to your TypoScript setup, the fallback images for
social media will only be rendered when no image is set in the page
properties. This will prevent that the fallback image is always
rendered even when an image is set in the backend.

Resolves: #93272
Releases: master, 10.4
Change-Id: I6efa046059f64d8dfc3d22bac9d451c9b402b81a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67594
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>