[BUGFIX] Re-allow rendering of IRRE subitems in Workspaces GridView
In previous TYPO3 versions (until the Workspace module was refactored
in TYPO3 v8) it was possible to manipulate the collapse / expand
toggle via Signals (now PSR-14 event), for third-party
records with a lot of IRRE records.

This change re-introduces the toggle for PSR-14 event listeners
to respond to the JavaScript toggle / flag.

Resolves: #93404
Releases: master, 10.4
Change-Id: I56408ac52b6491fe7ac79175ff0ab5017905d45d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67585
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>