[BUGFIX] Make ext:form previous step button work again
Since #92815 the extbase request dispatcher sets an "original
request" within a subrequest in case of a "ForwardResponse".

The fluid AbstractFormFieldViewHelper::hasMappingErrorOccurred()
guesses that a property mapping error exists as soon as a request
is a subrequest. This is determined by the fact that the current
request has a original request (e.g. if a controller returns a
ForwardResponse).

While the determination of "property mapping error exists" based
on the existence of a parent request may not be the best indicator,
it seems to be fine for EXT:form for now.

EXT:form makes further use of the HiddenViewHelper to set the previous
page index as a value for the property "__currentPage". Since this
ViewHelper always sets the respectSubmittedDataValue property to TRUE,
the value is always overwritten in case a parent request exists and
the "previous step" button did not longer work.

This is now fixed by making the respectSubmittedDataValue property
configurable via a ViewHelper argument and setting in to FALSE for
the "__currentPage" field.

Resolves: #93398
Releases: master
Change-Id: I8d495395c91ed8bcfae4fec4d474e73279e9e4ef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67590
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Benni Mack <benni@typo3.org>